Initial implementation of mob guard behavior
This commit is to start the new branch and for code review.

I've tried to make mobs guard their spawn points. In this initial revision
all mobs do this. The next step is to add a behavior field to mobs and spawns
to allow such extra AI behaviors to be specified by content developers.

Because the pathfinding code cannot find complex/long paths and I've no
idea how to fix it the guard behavior is implemented with breadcrumbs.
If the path code is ever made to be reliable we can undef BROKEN_PATHFIND
and remove the breadcrumb code.